tbo/requirements.txt
amadeus==12.0.0
redis==5.0.1
kafka-python==2.0.2
schedule==1.2.0
requests==2.31.0
qdrant-client==1.8.2
